**Summary**
Enable `qconv1d/2d/3d`, `qconv2d_relu`, `qconv2d_add`, and `qconv2d_add_relu` operator for quantization 2.0 export with oneDNN library.

**Test Plan**
```
python -u -m pytest -s -v test_quantized_op.py -k test_qconv1d_pt2e
python -u -m pytest -s -v test_quantized_op.py -k test_qconv2d_pt2e
python -u -m pytest -s -v test_quantized_op.py -k test_qconv3d_pt2e
python -u -m pytest -s -v test_quantized_op.py -k test_qconv2d_relu_pt2e
python -u -m pytest -s -v test_quantized_op.py -k test_qconv2d_add_pt2e
python -u -m pytest -s -v test_quantized_op.py -k test_qconv2d_add_relu_pt2e
```

cc jgong5 mingfeima XiaobingSuper sanchitintel ashokei jingxu10

[ghstack-poisoned]

+++ torch/testing/_internal/hypothesis_utils.py
@@ -189,7 +189,7 @@ Generates:
         (If `qparams` arg is None), returns None.
 """
 @st.composite
-def tensor(draw, shapes=None, elements=None, qparams=None, dtype=np.float32):
+def tensor(draw, shapes=None, elements=None, qparams=None):
     if isinstance(shapes, SearchStrategy):
         _shape = draw(shapes)
     else:
@@ -197,7 +197,7 @@ def tensor(draw, shapes=None, elements=None, qparams=None, dtype=np.float32):
     if qparams is None:
         if elements is None:
             elements = floats(-1e6, 1e6, allow_nan=False, width=32)
-        X = draw(stnp.arrays(dtype=dtype, elements=elements, shape=_shape))
+        X = draw(stnp.arrays(dtype=np.float32, elements=elements, shape=_shape))
         assume(not (np.isnan(X).any() or np.isinf(X).any()))
         return X, None
     qparams = draw(qparams)
@@ -205,7 +205,7 @@ def tensor(draw, shapes=None, elements=None, qparams=None, dtype=np.float32):
         min_value, max_value = _get_valid_min_max(qparams)
         elements = floats(min_value, max_value, allow_infinity=False,
                           allow_nan=False, width=32)
-    X = draw(stnp.arrays(dtype=dtype, elements=elements, shape=_shape))
+    X = draw(stnp.arrays(dtype=np.float32, elements=elements, shape=_shape))
     # Recompute the scale and zero_points according to the X statistics.
     scale, zp = _calculate_dynamic_qparams(X, qparams[2])
     enforced_zp = _ENFORCED_ZERO_POINT.get(qparams[2], None)
